+++ http/http-advanced-reactive/src/test/resources/test.properties
@@ -1,2 +1,2 @@
-ts.keycloak.log.enable=false
+ts.keycloak.log.enable=true
 ts.app.log.enable=true
